+++ src/mdp/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+1.0.0b2
+++++++
+* Upgrading to latest api version 2024-04-04-preview.
+
 1.0.0b1
 ++++++
 * Initial release.